# fix: Improve consistency between `read_csv` and `scan_csv`