fixes #22395

This commit improves consistency between `read_csv()` and `scan_csv()` for some types of malformed CSV data. It will issue a warning for some types of malformed CSV data.  A decision is needed on the scope as this may impact performance, see below.

In scope:
a) (Cheap) Return warning instead of panic when `read_chunk()` returns more lines than `CountLines` found. Add a similar warning to `scan_csv()`. This makes `read_csv` consistent with `scan_csv`. We may want to error instead?
b) Updated documentation.
c) Added tests.

TBD in/out of scope (currently in):
d) (Not cheap) Lazy_quotes_1: Looks for quotes unenclosed fields in `SplitFields` iterator: detect and warn when quotes are found in unenclosed field, except for columns to the RHS of last projected column.

Out of scope:
e) Allow lazy_quotes_1 with projection, specifically when columns to the RHS of the last projected column have an odd number of quotes.
f) Lazy_quotes_2: Does not check for single-quotes in enclosed fields (yet) as this is more ambiguous and further impacts performance. Exception is the line count above.
g) Leading whitespace in fields.

Performance regression (for unenclosed fields lazy_quotes_1 check, d above) observed. Caveat, my dev machine is not good for benchmarking, it is a lightweight portable laptop running Windows + WSL. It shows some performance regression on a single large CSV file (NYC dataset, 24M records, short fields, unenclosed) in the order of +9% on number of instructions, +15% on wall-time.

<details>
<summary>Details: (non-representative) performance assessment</summary>

(with malformed check)
(nodebug-release)
commit 1d60b0488b29fe6c7d4292a7f49da6e1f025ca23 

```
denecker@DESKTOP-V7KT63E:~/my-polars/issue_22395$ ~/poop/zig-out/bin/poop 'python nyc_read.py'
Benchmark 1 (3 runs): python nyc_read.py
  measurement          mean ± σ            min … max           outliers
  wall_time          6.14s  ±  723ms    5.47s  … 6.91s           0 ( 0%)
  peak_rss           8.43GB ± 6.33MB    8.43GB … 8.44GB          0 ( 0%)
  cpu_cycles          102G  ±  487M      101G  …  102G           0 ( 0%)
  instructions        141G  ±  395K      141G  …  141G           0 ( 0%)
  cache_references    561M  ± 34.0M      541M  …  600M           0 ( 0%)
  cache_misses        333M  ± 4.69M      327M  …  336M           0 ( 0%)
  branch_misses       306M  ± 9.31M      298M  …  316M           0 ( 0%)
```

(without malformed check)
(nodebug-release)
commit 5be0a9b21cd84f0a9f6c3e159c8668584d81608e

```
denecker@DESKTOP-V7KT63E:~/my-polars/issue_22395$ ~/poop/zig-out/bin/poop 'python nyc_read.py'
Benchmark 1 (3 runs): python nyc_read.py
  measurement          mean ± σ            min … max           outliers
  wall_time          5.33s  ±  169ms    5.22s  … 5.53s           0 ( 0%)
  peak_rss           8.43GB ±  967KB    8.43GB … 8.44GB          0 ( 0%)
  cpu_cycles         90.0G  ±  332M     89.7G  … 90.4G           0 ( 0%)
  instructions        129G  ±  463K      129G  …  129G           0 ( 0%)
  cache_references    540M  ± 17.3M      526M  …  560M           0 ( 0%)
  cache_misses        331M  ± 1.67M      329M  …  333M           0 ( 0%)
  branch_misses       288M  ± 14.1M      280M  …  305M           0 ( 0%)
```
</details>

General notes on performance. I suspect overall, some improvements are still possible for `read_csv` if we consider the following:
(A) Internal
1. Use bounded queue between pass-1 and pass-2 to provide back-pressure. This should keep passes in sync and keep cache hot (requires investigation).
2. Review L3 cache size optimization (individual chunks are 16 MB, per thread?)
3. In pass 2, consider collecting character indices for further processing rather than re-starting 'next()` for every field. Relevant when fields are short.
4. I noticed SIMD_SIZE is hardcoded as 64 bytes (AVX-512) but many Intel laptops only support 32 bytes (AVX-2). (Fwiw, hardcoding SIMD_SIZE to 32 bytes showed some improvement on my machine).

(B) User facing:
1. Introduce a `fast-mode` option to end users in `read_csv()`. When enabled, polars assert wellformed data  and ignores all quotes.
2. Introduce feature flag(s) for lazyquotes options with separate code path. This protects the performance on wellformed CSV. Consider memchr to manage code complexity.
Combined, this would present 3 approaches: (i) fast-mode (expects well-formed, no quotes, no checks), (ii) default mode (expects well-formed including quotes, limited checking), (iii) forgiving mode (feature flags to allow RFC exceptions).

Kindly review:
- Issue `warning` vs `error`. My take is to start with `warning` now and move to `error` when forgiving feature flags are available.
- Scope: how much performance impact is acceptable to detect malformed data? Imo, the next step would be a reliable benchmark on a dedicated machine.