@kdn36 2025-05-23
Sharing WIP.
Open items: 
- Strict line count check when there are comments in a chunk.
- Consistency when `select(pl.nth())` is used.